rules_r: output ignored package files
Summary:
This adds an output group to output files that contain the diff between
the file list contained by the binary archive of the R package, and the
files we instruct bazel to collect.

This will help close the gap between an `R CMD INSTALL` install and our
bazel install.

These files can be generated by:
  bazel build @R_bitops//... --output_groups=pkg_file_list_diffs

And accessed by:
  bazel-bin/external/R_bitops/pkg_file_list_diff.txt

Reviewers: razvanm

Reviewed By: razvanm

Differential Revision: https://phabricator.grailbio.com/D12647

fbshipit-source-id: 62f9d46