rules_r: reproducible builds on demand
Summary:
Best effort feature to produce files with identical checksum on
successive builds. Activated with --features=rlang-reproducible.

These are slightly more expensive and are more complex because of the
need to have the same src and install paths. And depends on the
compiler's willingness to produce reproducible .so files.

Known to not work on macOS where the .o files can be at different
positions within the .so, and on packages with C++ cc_deps.

Reviewers: razvanm

Reviewed By: razvanm

Differential Revision: https://phabricator.grailbio.com/D12645

fbshipit-source-id: fadd8e5