rules_r: fail lazily in r_repository_list
Summary:
Since the generated bzl file is loaded in the workspace, everyone needs
to process it regardless of whether they are using R packages or not.

Now bazel will fail only when it _attempts_ to use a non-existent
package.

Test Plan:
Tested by mistyping a package name in the package list. All builds
worked fine except for the building the remote package.

Reviewers: ysaito

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D12889

fbshipit-source-id: b642e9f