rules_r: use package lib as declared directory
Summary:
This gets rid of individual package file declarations and rule
attributes needed to guess which files will be generated.

The built file set is now exactly the same as `R CMD INSTALL`,
and the code is much simpler.

This change comes now in the light of new knowledge that declared
directories also include files (recursively) underneath them.

Reviewers: ysaito

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D14397

fbshipit-source-id: 560e175