rules_r: make reproducible builds the default
Summary:
GRAIL has been building with the feature enabled for many months and
has not faced a problem. Making this the default and removing the other
option simplifies the code.

As a bonus to this change, we also make our source and binary package
builds consistent.

Test Plan: Unit tests

Reviewers: ysaito

Reviewed By: ysaito

Maniphest Tasks: T12955

Differential Revision: https://phabricator.grailbio.com/D24996

fbshipit-source-id: 6c980d5