rules_r: speedup installs of R libraries
Summary:
This commit adds the --checksum option to the rsync command used to
install R libraries into LIBRARY_PATH after a bazel build. Without this
option, rsync uses timestamps and file sizes to decide what needs to be
copied, which are not always reliable.

This should not be a problem usually on laptops with SSDs
but security scanners can scan all writes and cause bottlenecks.

Test Plan: Tested local builds.

Reviewers: sbagaria, ysaito

Reviewed By: sbagaria, ysaito

Differential Revision: https://phabricator.grailbio.com/D29463

fbshipit-source-id: 3fe31a0