rules_r: new r_markdown rule
Summary:
This rule provides an executable target, which when run with `bazel
run`, will render the given input file through the given render
function (default is rmarkdown::render).

If an argument is given on the command line, it will be the output
directory, else the output directory will be the default output
directory of the render function, typically the same directory as the
input file.

This rule can be used with the `r_binary_image` rule.

Test Plan: Tested a hello world Rmd with this rule.

Reviewers: dfilippova, jkim

Reviewed By: jkim

Maniphest Tasks: T12955

Differential Revision: https://phabricator.grailbio.com/D29439

fbshipit-source-id: c8ffb00