rules_r: add new R and Bioc releases to repo management script
Summary:
We have checks in place to compute if the constructed binary repo URL is
valid or not. These checks are hardcoded to specific R and Bioc
versions, and so have to be manually updated.

Test Plan:
Tested updating the BiocParallel package; got a valid pre-compiled
package for R 3.5 on Bioc 3.8.

Reviewers: ccosenza

Reviewed By: ccosenza

Maniphest Tasks: T12955

Differential Revision: https://phabricator.grailbio.com/D29967

fbshipit-source-id: 7f2eb9a