rules_r: cmdline args passing to r_markdown
Summary:
This allows to pass flags in the form of --arg=value where arg is the
name of the argument to the render function, and value is the R
expression for the argument.

Test Plan:
Tested rendering markdowns with the params argument given on the command
line to `bazel run`.

Reviewers: jkim

Reviewed By: jkim

Maniphest Tasks: T12955

Differential Revision: https://phabricator.grailbio.com/D30379

fbshipit-source-id: de1e5f7