rules_r: improve rsync options
Summary:
The current rsync options were inherting the (read-only) permissions
from the source files. And so we had to change the permissions
separately later.

We now provide the right set of options to rsync to mimic the --archive
flag in ways that we want, and be different from it in ways that we do
not want.

Test Plan:
Tested installing packages with these options. There were no
performance degradations.

Reviewers: ccosenza, schandra, ysaito

Reviewed By: ccosenza, schandra, ysaito

Maniphest Tasks: T12955

Differential Revision: https://phabricator.grailbio.com/D30619

fbshipit-source-id: 80bc4de