rules_r: specify macOS binary packages for coverage deps
This helps the tests on macOS where the dependencies needed to build
these packages are not always available easily. At GRAIL, we have not
depended on the coverage deps because we include the deps of covr in
our own repository list.

NOTE: This is a breaking change for people who depended on
r_coverage_dependencies as its usage now includes two extra lines. The
README has been changed to reflect the updated instructions.

As an extension, we are also updating the dependencies for the tests of
this repo and enabling the use of binary packages where available.