rules_r: stamp attribute for r_binary and r_test
This stamp attribute will determine whether we depend on the stable
status file in the runfiles of the binary/test. The volatile status file
is always available. This will help people get better cache hits across
builds when their workspace status command is outputting STABLE_ vars
but an individual test does not really need these vars.

NOTE: This is a breaking change for people who were depending on the
stable status file in their r_test or r_markdown target. This file will
not be available by default because the attribute value is False by
default. Volatile status files will continue to be available.

See #37 for more context.

We expect the usage of this attribute to be limited, and we expect
cache hits in stamping to be governed mostly through the management of
the content of stable status file.

The logic extends to r_markdown.

Co-authored-by: forestfang-stripe <45770568+forestfang-stripe@users.noreply.github.com>