Fix building vignettes
Building vignettes only work with the `R CMD build` command and not with
the `R CMD INSTALL` command, which we had previously assumed.

This change ensures that we "install" from the source files that are
built using the `R CMD build` command. It may slow down our build
process.

Also adds tests that vignettes are getting built (needs texinfo).