Cleanup files in tests after R CMD build
The .o files in the src/lib directory of exampleC package are
non-standard and are not automatically cleaned up by `R CMD BUILD`. So
they need to be cleaned manually.

exampleC/src/lib/*.o also contains the compile dir which is a temp
directory and makes our reproducibility test fail because exampleC.so is
not reproducible.