Allow package sources at workspace root
The codebase had an assumption that the bazel package path for the r_pkg
target is non-empty. We remove that assumption and make sure that we
don't include special directories at the execroot like bazel-out and
external in the package source.

Bonus: Also makes state_test not rely on a hard-coded version of R. This
allows easier testing on local machines.

Fixes #49.