Mark repository_list as system configurable
This allows one to reset the repository list after installing or
upgrading R to a different version.

Also, take Rscript as a parameter to users can specify the same Rscript
here as the toolchain they are using. This is because we can not use the
resolved toolchain directly in a repository rule.