Set PKG_ vars for cc_deps in the site Makevars file (#65)
Package src/Makevars files typically use the `=` assignment because `+=` and `:=` are both GNU extensions. Which means that the environment variable values are overwritten by the user-defined list in the package src/Makevars.

By using the site Makevars file with the `+=` assignment, we can ensure that we append to any user-specified list in src/Makevars, and that our values are not overwritten by the user-specified list. Because site Makevars file is processed after the package src/Makevars file.

Fixes #63.