Separate source and binary build steps (#73)
When building a package, we make the source archive (without tests)
first. If the user requests both the source and the binary archives, we
have been duplicating work so far.

This change breaks apart the two steps and introduces a dependency
between the two. This will avoid any duplicate work.

As part of this refactoring, we also resolved a TODO to make these tars
reproducible. This way, the intermediate source archive is reproducible.

Fixes #69.