Create darwin Makevars when gfortran is unavailable
We were previously not creating the Makevars at all when gfortran was
unavailable because so many packages depend on gfortran. However, we
have since moved to using pre-built binary packages on macOS, which
means that in most projects, users would be able to get by without
gfortran because they would need to compile very few external packages
from source.