Reproducibility of native code on macOS
LLVM versions on macOS have been updated to be past LLVM 7, so
-fdebug-prefix-map now works as expected and the reproducibility tests
can be re-enabled.

Now, we just need to set ZERO_AR_DATE=1 (thanks @chrchang).
Also see https://github.com/bazelbuild/bazel/issues/10886

There is still a problem with the version of `install_name_tool`
shipped with Xcode CLT, and therefore we are starting to distribute a
newer version of `install_name_tool` in this project.