Better reproducibility of native code
We notice that in repro tests run locally, we were using a remote cache,
and so were not properly testing for reproducibility.

After fixing the test, we discovered that were not using
`-fno-canonical-system-headers` for gcc which was embedding paths of
cc_deps header directories in compiled .so files. When using headers
from R's "Linking To" type dependencies, the library directory where the
directory is installed needs to be a constant path, or the path needs to
be added to -fdebug-prefix-map options.