Document reproduciblity flags in the README
BREAKING CHANGE: Users defining their own toolchains will now need to
ensure that they are defining the right compiler flags for
reproduciblity in their site Makevars file. They are no longer a part
of the package build script.

We moved the compiler flags needed for reproducible builds from the
build script to the site Makevars files in the default local toolchains
in the previous commit.