Set up Buildbuddy remote cache (#79)
Buildbuddy is authenticated as per flags in tests/auth.bazelrc. If file is
absent, the build is uploaded unauthenticated. On Github actions,
this file is created if the relevant secrets are present.

If remote_cache is set up in homerc or systemrc files, then that is
given preference.

Bonus changes:
- Remove setting up an explicit version of R; latest version is good.
- Tried to make 3.5.0 work with tests, but an additional system include
   directory on macOS is harder than what I want.
- Use CPLUS_INCLUDE_PATH instead of symlinking.