Use file roots to construct paths for runfiles
We were previously using the fixed bin_dir (and the now deprecated
gen_dir) constants to construct runfile paths. But the fixed bin_dir
does not work for configuration transitions. bazel now provides a
File.root field to determine if the file lives under a file root, which
could be a variable bin_dir depending on the transition in effect.