Rethink handling of status files
It appears the --stamp flag has had wider adoption, and along with
more use of remote caching which treats the volatile status file
differently, having a simpler flag to turn the inclusion of status files
on/off is easier for the users. Note that the choice to respect the
--stamp flag is on each rule author in general, and bazel does not
guarantee any behavior with the flag.

For including status files in `r_binary` type rules, the users would
have to explicitly include them in as data dependencies.