Fix runfiles handling
Runfiles handling in rules_r was an afterthought and we were dropping
runfiles from some dependencies in various places. This change makes an
attempt to go over all the flows and capture runfiles appropriately.

Note that we are currently not using the runfiles from cc_deps, but it
can change if we want.